FlxBitmapText fix, tweaks
* fixed weird FlxBitmapText issue - looks like the font wasn't being
reinstantiated or something? maybe the fonts are pooled and if they
share the same graphic it doesn't create a new one?? To not make the
same mistake the font created in UI now references the same alphabet as
the ones in the PlayState (PlayState.large_alphabet)
* Stars indicate new hi-score on gameover screen - not super WOW, could
be better, use something else
* combos now trigger and wait until bar is at 50% before disabling the
boost ability - could probably be hiked up to 65% at the most, but this
gives players some flexibility  and creates a smoother experience.